requirements.txt
mesop==0.12.2
requests
google_generativeai
gunicorn
